Authentication for Api added...

+++ CouponsWorld/src/main/java/com/couponsworld/resources/OfferResource.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -18,8 +19,10 @@ import com.couponsworld.dto.Link;
 import com.couponsworld.dto.Offer;
 import com.couponsworld.dto.ResultantOffer;
 import com.couponsworld.enums.Status;
+import com.couponsworld.services.GenerateLinkService;
 import com.couponsworld.services.OfferService;
-import com.google.appengine.api.search.query.QueryParser.restriction_return;
+import com.couponsworld.utilities.AuthenticationKeyValidator;
+import com.couponsworld.utilities.Constants;
 
 @Path("/offers")
 public class OfferResource {
@@ -29,10 +32,54 @@ public class OfferResource {
 	private ResultantOffer resultantOffer;
 
 	@GET
+	@Path("/{authLoginId}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public ResultantOffer getOffers() {
+	public ResultantOffer getOffers(@PathParam("authLoginId") String authLoginId,
+			@HeaderParam("authKey") String authKey) {
 		try {
-			return OfferService.getOffers();
+			if (authLoginId.equals(Constants.AUTH_LOGIN_ID) && AuthenticationKeyValidator.validate(authKey)) {
+				return OfferService.getOffers();
+			} else {
+				// creating resultantOffer Object
+				resultantOffer = new ResultantOffer();
+
+				// creatin g the error getting
+				com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+				error.setErrorCode(102);
+				error.setErrorName("Invalid Login Credentials....");
+
+				// wrapping the error to a list of errors
+				errors = new ArrayList<Error>();
+				errors.add(error);
+
+				resultantOffer.setErrors(errors);
+				resultantOffer.setLinks(GenerateLinkService.generateLink("getOffers"));
+				resultantOffer.setStatus(Status.FAILURE);
+				resultantOffer.setOffers(offers);
+				errors = null;
+				offers = null;
+				return resultantOffer;
+			}
+
+		} catch (NullPointerException npe) {
+			// creating resultantOffer Object
+			resultantOffer = new ResultantOffer();
+
+			// creating the error getting
+			com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+			error.setErrorCode(101);
+			error.setErrorName("Provide correct credentials to access api");
+
+			// wrapping the error to a list of errors
+			errors = new ArrayList<Error>();
+			errors.add(error);
+
+			resultantOffer.setErrors(errors);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("getOffers"));
+			resultantOffer.setStatus(Status.FAILURE);
+			resultantOffer.setOffers(offers);
+			errors = null;
+			return resultantOffer;
 		} catch (Exception exception) {
 
 			// creating resultantOffer Object
@@ -41,27 +88,73 @@ public class OfferResource {
 			// creating the error getting
 			com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
 			error.setErrorCode(101);
-			error.setErrorName(exception.getMessage().toString());
+			error.setErrorName(exception.getMessage());
 
 			// wrapping the error to a list of errors
 			errors = new ArrayList<Error>();
 			errors.add(error);
 
 			resultantOffer.setErrors(errors);
-			resultantOffer.setLinks(links);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("getOffers"));
 			resultantOffer.setStatus(Status.FAILURE);
 			resultantOffer.setOffers(offers);
+			errors = null;
 			return resultantOffer;
 		}
 
 	}
 
 	@POST
+	@Path("/{authLoginId}")
 	@Produces(MediaType.APPLICATION_JSON)
 	@Consumes(MediaType.APPLICATION_JSON)
-	public ResultantOffer createOffer(Offer offer) {
+	public ResultantOffer createOffer(Offer offer, @PathParam("authLoginId") String authLoginId,
+			@HeaderParam("authKey") String authKey) {
 		try {
-			return OfferService.createOffer(offer);
+			if (authLoginId.equals(Constants.AUTH_LOGIN_ID) && AuthenticationKeyValidator.validate(authKey)) {
+				return OfferService.createOffer(offer);
+			} else {
+				// creating resultantOffer Object
+				resultantOffer = new ResultantOffer();
+
+				// creatin g the error getting
+				com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+				error.setErrorCode(102);
+				error.setErrorName("Invalid Login Credentials....");
+
+				// wrapping the error to a list of errors
+				errors = new ArrayList<Error>();
+				errors.add(error);
+
+				resultantOffer.setErrors(errors);
+				resultantOffer.setLinks(GenerateLinkService.generateLink("createOffer"));
+				resultantOffer.setStatus(Status.FAILURE);
+				resultantOffer.setOffers(offers);
+
+				offers = null;
+				errors = null;
+				return resultantOffer;
+			}
+		} catch (NullPointerException npe) {
+			// creating resultantOffer Object
+			resultantOffer = new ResultantOffer();
+
+			// creating the error getting
+			com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+			error.setErrorCode(101);
+			error.setErrorName("Provide correct credentials to access api");
+
+			// wrapping the error to a list of errors
+			errors = new ArrayList<Error>();
+			errors.add(error);
+
+			resultantOffer.setErrors(errors);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("createOffer"));
+			resultantOffer.setStatus(Status.FAILURE);
+			resultantOffer.setOffers(offers);
+
+			errors = null;
+			return resultantOffer;
 		} catch (Exception exception) {
 
 			// creating Offer List to wrap the input Offer Object into it
@@ -81,20 +174,65 @@ public class OfferResource {
 			errors.add(error);
 
 			resultantOffer.setErrors(errors);
-			resultantOffer.setLinks(links);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("createOffer"));
 			resultantOffer.setStatus(Status.FAILURE);
 			resultantOffer.setOffers(offers);
+			errors = null;
 			return resultantOffer;
 		}
 	}
 
 	@PUT
-	@Path("/{offerId}")
+	@Path("/{offerId}/{authLoginId}")
 	@Produces(MediaType.APPLICATION_JSON)
 	@Consumes(MediaType.APPLICATION_JSON)
-	public ResultantOffer updateOffer(@PathParam("offerId") long offerId, Offer offer) {
+	public ResultantOffer updateOffer(@PathParam("offerId") long offerId, Offer offer,
+			@PathParam("authLoginId") String authLoginId, @HeaderParam("authKey") String authKey) {
 		try {
-			return OfferService.updateOffer(offerId, offer);
+			if (authLoginId.equals(Constants.AUTH_LOGIN_ID) && AuthenticationKeyValidator.validate(authKey)) {
+				return OfferService.updateOffer(offerId, offer);
+			} else {
+				// creating resultantOffer Object
+				resultantOffer = new ResultantOffer();
+
+				// creatin g the error getting
+				com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+				error.setErrorCode(102);
+				error.setErrorName("Invalid Login Credentials....");
+
+				// wrapping the error to a list of errors
+				errors = new ArrayList<Error>();
+				errors.add(error);
+
+				resultantOffer.setErrors(errors);
+				resultantOffer.setLinks(GenerateLinkService.generateLink("updateOffer"));
+				resultantOffer.setStatus(Status.FAILURE);
+				resultantOffer.setOffers(offers);
+
+				offers = null;
+				errors = null;
+				return resultantOffer;
+			}
+		} catch (NullPointerException npe) {
+			// creating resultantOffer Object
+			resultantOffer = new ResultantOffer();
+
+			// creating the error getting
+			com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+			error.setErrorCode(101);
+			error.setErrorName("Provide correct credentials to access api");
+
+			// wrapping the error to a list of errors
+			errors = new ArrayList<Error>();
+			errors.add(error);
+
+			resultantOffer.setErrors(errors);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("updateOffer"));
+			resultantOffer.setStatus(Status.FAILURE);
+			resultantOffer.setOffers(offers);
+
+			errors = null;
+			return resultantOffer;
 		} catch (Exception exception) {
 
 			// creating Offer List to wrap the input Offer Object into it
@@ -114,19 +252,65 @@ public class OfferResource {
 			errors.add(error);
 
 			resultantOffer.setErrors(errors);
-			resultantOffer.setLinks(links);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("updateOffer"));
 			resultantOffer.setStatus(Status.FAILURE);
 			resultantOffer.setOffers(offers);
+
+			errors = null;
 			return resultantOffer;
 		}
 	}
 
-	@Path("/{offerId}")
 	@DELETE
+	@Path("/{offerId}/{authLoginId}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public ResultantOffer deleteOffer(@PathParam("offerId") long offerId) {
+	public ResultantOffer deleteOffer(@PathParam("offerId") long offerId, @PathParam("authLoginId") String authLoginId,
+			@HeaderParam("authKey") String authKey) {
 		try {
-			return OfferService.deleteOffer(offerId);
+			if (authLoginId.equals(Constants.AUTH_LOGIN_ID) && AuthenticationKeyValidator.validate(authKey)) {
+				return OfferService.deleteOffer(offerId);
+			} else {
+				// creating resultantOffer Object
+				resultantOffer = new ResultantOffer();
+
+				// creatin g the error getting
+				com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+				error.setErrorCode(102);
+				error.setErrorName("Invalid Login Credentials....");
+
+				// wrapping the error to a list of errors
+				errors = new ArrayList<Error>();
+				errors.add(error);
+
+				resultantOffer.setErrors(errors);
+				resultantOffer.setLinks(GenerateLinkService.generateLink("deleteOffer"));
+				resultantOffer.setStatus(Status.FAILURE);
+				resultantOffer.setOffers(offers);
+
+				offers = null;
+				errors = null;
+				return resultantOffer;
+			}
+		} catch (NullPointerException npe) {
+			// creating resultantOffer Object
+			resultantOffer = new ResultantOffer();
+
+			// creating the error getting
+			com.couponsworld.dto.Error error = new com.couponsworld.dto.Error();
+			error.setErrorCode(101);
+			error.setErrorName("Provide correct credentials to access api");
+
+			// wrapping the error to a list of errors
+			errors = new ArrayList<Error>();
+			errors.add(error);
+
+			resultantOffer.setErrors(errors);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("deleteOffer"));
+			resultantOffer.setStatus(Status.FAILURE);
+			resultantOffer.setOffers(offers);
+
+			errors = null;
+			return resultantOffer;
 		} catch (Exception exception) {
 
 			// creating resultantOffer Object
@@ -142,9 +326,11 @@ public class OfferResource {
 			errors.add(error);
 
 			resultantOffer.setErrors(errors);
-			resultantOffer.setLinks(links);
+			resultantOffer.setLinks(GenerateLinkService.generateLink("deleteOffer"));
 			resultantOffer.setStatus(Status.FAILURE);
 			resultantOffer.setOffers(offers);
+
+			errors = null;
 			return resultantOffer;
 		}
